CODE RUB: Update HomePage
