+++ LondonDataServices.IDecide.Portal.Server/Controllers/PatientsController.cs
@@ -67,12 +67,13 @@ namespace LondonDataServices.IDecide.Portal.Server.Controllers
             var patient = new Patient
             {
                 Id = id,
-                NhsNumber = "9876543210",
-                FirstName = "Jane",
-                Surname = "Smith",
-                EmailAddress = "jane.smith@example.com",
-                Address = "456 High St",
-                Postcode = "CD34 5EF",
+                NhsNumber = "1234567890",
+               FirstName = "D****",
+                Surname = "H****",
+                EmailAddress = "d********s@googlemail.com",
+                PhoneNumber = "07*********",
+                Address = "9 The Wood******, S**********, Surrey",
+                Postcode = "CR2 0HG",
                 DateOfBirth = new DateTime(1990, 5, 15),
                 CreatedBy = "system",
                 CreatedDate = DateTime.UtcNow,
@@ -90,12 +91,12 @@ namespace LondonDataServices.IDecide.Portal.Server.Controllers
             {
                 Id = Guid.NewGuid().ToString(),
                 NhsNumber = patient?.NhsNumber ?? "N/A",
-                FirstName = "Jane",
-                Surname = "Smith",
-                EmailAddress = "jane.smith@example.com",
-                PhoneNumber = patient?.PhoneNumber ?? "0000000000",
-                Address = "456 High St",
-                Postcode = "CD34 5EF",
+                FirstName = "D****",
+                Surname = "H****",
+                EmailAddress = "d********s@googlemail.com",
+                PhoneNumber = "07*********",
+                Address = "9 The Wood******, S**********, Surrey",
+                Postcode = "CR2 0HG",
                 DateOfBirth = patient?.DateOfBirth == default ? new DateTime(1990, 5, 15) : patient.DateOfBirth,
                 CreatedBy = "system",
                 CreatedDate = DateTime.UtcNow,
